# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant video platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos/{videoId}/thumbnail-upload-intent:
    post:
      operationId: createThumbnailUploadIntent
      tags: [videos]
      parameters:
        - $ref: "#/components/parameters/VideoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailUploadIntentInput"
      responses:
        "201":
          description: Short-lived exact-object upload capability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThumbnailUploadIntent"
  /videos/{videoId}/thumbnail-finalize:
    post:
      operationId: finalizeThumbnail
      tags: [videos]
      parameters:
        - $ref: "#/components/parameters/VideoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailFinalizeInput"
      responses:
        "200":
          description: Finalized thumbnail application route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thumbnail"
  /videos/{videoId}/thumbnail:
    get:
      operationId: getVideoThumbnail
      tags: [videos]
      parameters:
        - $ref: "#/components/parameters/VideoId"
      description: Requires the opaque current `v` query value returned in thumbnailUrl.
      responses:
        "200":
          description: Authenticated thumbnail bytes
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
    delete:
      operationId: deleteVideoThumbnail
      tags: [videos]
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "204":
          description: Thumbnail removed
...
components:
  parameters:
    VideoId:
      in: path
      name: videoId
      required: true
      schema:
        type: string
...
  schemas:
...
    ThumbnailUploadIntentInput:
      type: object
      additionalProperties: false
      required: [contentType, sizeBytes]
      properties:
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sizeBytes: { type: number, minimum: 1, maximum: 10485760 }
    ThumbnailUploadIntent:
      type: object
      required: [intentId, uploadUrl, requiredHeaders, expiresAt]
      properties:
        intentId: { type: string }
        uploadUrl: { type: string }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
        expiresAt: { type: string, format: date-time }
    ThumbnailFinalizeInput:
      type: object
      additionalProperties: false
      required: [intentId]
      properties:
        intentId: { type: string }
    Thumbnail:
      type: object
      required: [thumbnailUrl, contentType, sizeBytes]
      properties:
        thumbnailUrl:
          type: string
          description: Versioned owned application URL. The object key and storage generation are never exposed.
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sizeBytes: { type: number, minimum: 1, maximum: 10485760 }
...